requirements.txt
langchain
langchain-openai
langchain-community
openai
pydantic
python-dotenv
pypdf
tiktoken
langgraph
python-docx
docx2txt
chromadb
sentence-transformers